[consumer] Refactor the consumer
By looping over events, we can keep the consumer alive even when
messages aren't coming in.

Also had to fix the timestmap model since we should be getting
timestamps with 'Z' in them. If we already have that, we were getting
errors by trying to add a 'Z' to the end.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>